requirements.txt
mal-api==0.5.2

requests==2.28.1
lxml==4.9.1
bs4==0.0.1
openai